Trying 100 times to give back function values, and only then killing

+++ libensemble/gen_funcs/aposmm_localopt_support.py
@@ -160,19 +160,26 @@ class LocalOptInterfacer(object):
 
     def destroy(self, previous_x):
 
-        # while not isinstance(previous_x, ConvergedMsg):
-        #     self.parent_can_read.clear()
-        #     if self.grad0 is None:
-        #         self.comm_queue.put((previous_x, 0*np.ones_like(self.f0),))
-        #     else:
-        #         self.comm_queue.put((previous_x, 0*np.ones_like(self.f0), np.zeros_like(self.grad0)))
+        count = 0
+        while not isinstance(previous_x, ConvergedMsg) and count <= 100:
+            self.parent_can_read.clear()
+            if self.grad0 is None:
+                self.comm_queue.put((previous_x, 0*np.ones_like(self.f0),))
+            else:
+                self.comm_queue.put((previous_x, 0*np.ones_like(self.f0), np.zeros_like(self.grad0)))
 
-        #     self.child_can_read.set()
-        #     self.parent_can_read.wait()
+            self.child_can_read.set()
+            self.parent_can_read.wait()
 
-        #     previous_x = self.comm_queue.get()
-        # assert isinstance(previous_x, ConvergedMsg)
-        self.process.terminate()
+            previous_x = self.comm_queue.get()
+            count += 1
+
+        if not isinstance(previous_x, ConvergedMsg):
+            if hasattr(self.process, 'kill'):
+                self.process.kill()
+            else: 
+                self.process.terminate()
+
         self.process.join()
         self.comm_queue.close()
         self.comm_queue.join_thread()
